requirements.txt
# Dependencias del Agente de Automatización SQL

mysql-connector-python
PyMySQL
pandas
numpy
matplotlib
seaborn
plotly
jinja2
schedule
psutil
secure-smtplib
python-dotenv
colorama
rich
loguru
pyyaml
configparser
python-dateutil
pydantic
click
typer
Flask